requirements.txt
flask
langchain
langchain-community
langchain-core
langchain-groq
langchain-google-genai
google-generativeai
python-dotenv